supports chainging the configuration in real time

+++ ev3_software/KarolK72.LegoAssignment.Library/Commands/CommandAttribute.cs
@@ -8,9 +8,9 @@ namespace KarolK72.LegoAssignment.Library.Commands
 {
     public class CommandAttribute : Attribute
     {
-        private int _commandID = 0;
-        public int CommandID { get { return _commandID; } }
-        public CommandAttribute(int commandID)
+        private UInt16 _commandID = 0;
+        public UInt16 CommandID { get { return _commandID; } }
+        public CommandAttribute(UInt16 commandID)
         {
             _commandID = commandID;
         }
